переделал layout с Outlet на children.
