+++ src/app/app.tsx
@@ -1,5 +1,5 @@
 import { RouterProvider } from 'react-router-dom';
-import router from '../routes/router';
+import router from '../router/router';
 
 export default function App(): JSX.Element {
   return <RouterProvider router={router} />;
